fix: #52 리스트 댓글 삭제 본인만 가능하게 컨트롤러 수정

+++ matzipback/src/main/java/com/matzip/matzipback/matzipList/command/application/controller/ListCmtCommandController.java
@@ -49,7 +49,9 @@ public class ListCmtCommandController {
     public ResponseEntity<SuccessResMessage> deleteListCmt(@Valid @RequestBody DeleteListCmtRequset deleteListCmtRequset){
 
         try { if (CustomUserUtils.getCurrentUserAuthorities().iterator().next().getAuthority().equals("user")) {
-
+            if (!CustomUserUtils.getCurrentUserSeq().equals(deleteListCmtRequset.getListCommentUserSeq())) {
+                throw new RestApiException(FORBIDDEN_ACCESS);
+            }
             listCmtCommandService.deleteListCmt(deleteListCmtRequset);
 
             return ResponseEntity
